requirements.txt
# Core ML libraries
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
scipy>=1.7.0

# Diffusion and transformers
diffusers>=0.21.0
transformers>=4.21.0
accelerate>=0.20.0

# Data processing
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-image>=0.18.0

# Evaluation and metrics
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
wandb>=0.12.0  # optional, for experiment tracking
tensorboard>=2.8.0  # optional, for logging

# Configuration and I/O
pyyaml>=6.0
omegaconf>=2.1.0  # optional, for advanced config management

# Development and testing
pytest>=6.0.0  # for testing
black>=22.0.0  # for code formatting
isort>=5.10.0  # for import sorting

# Optional: for faster attention computation
# xformers>=0.0.16  # uncomment if available